Update cycletracks.py
"Trips" table in downloaded CSVs have "notes" as their last field - edited crusher to account for this.
First is that it wasn't identifying the "trips" table. Changed line 80 to look for 'notes'
Second is that it predictably freaks out whenever a note has a comma. Added a "max splits" (9) argument to the string split in the Trip line tuple.
Third edited things so that the notes get added to the Trips output table.

Still getting a few exceptions which appear to trace back to one user, userid=192, whose zip codes are a garbled mess. Added exception handling to function printToFile to deal with this.